ADD: default admin user (id: 10001) protection, double protections on Service and Controller layer

+++ src/main/java/cn/opencil/controller/authorization/AuthRoleController.java
@@ -35,6 +35,9 @@ public class AuthRoleController {
     @ResponseStatus(HttpStatus.CREATED)
     public RestfulResult grantPermissionToRole(@RequestBody JSONObject input) throws ValidationException, SimpleHttpException {
         RBACPermissionRole permissionRole = ValidationUtils.validate(input.toJavaObject(RBACPermissionRole.class), PermissionRoleIdVaildation.class);
+        if (permissionRole.getRoleId().equals(1)) {
+            throw new SimpleHttpException(400, "the default admin user can not be deleted", HttpStatus.BAD_REQUEST);
+        }
         if (!permissionRoleService.grantPermissionToRole(permissionRole)) {
             throw new SimpleHttpException(500, "database access error", HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -48,6 +51,9 @@ public class AuthRoleController {
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void revokePermissionFromRole(@RequestBody JSONObject input) throws ValidationException, SimpleHttpException {
         RBACPermissionRole permissionRole = ValidationUtils.validate(input.toJavaObject(RBACPermissionRole.class), PermissionRoleIdVaildation.class);
+        if (permissionRole.getRoleId().equals(1)) {
+            throw new SimpleHttpException(400, "the default admin user can not be deleted", HttpStatus.BAD_REQUEST);
+        }
         if (!permissionRoleService.revokePermissionFromRole(permissionRole)) {
             throw new SimpleHttpException(500, "database access error", HttpStatus.INTERNAL_SERVER_ERROR);
         }
@@ -61,7 +67,7 @@ public class AuthRoleController {
     public RestfulResult addRole(@RequestBody JSONObject input) throws ValidationException, SimpleHttpException {
         RBACPermissionRole permissionRole = ValidationUtils.validate(input.toJavaObject(RBACPermissionRole.class), RoleOperationValidation.class);
         if (!permissionRoleService.addRole(permissionRole)) {
-            throw new SimpleHttpException(500, "database access error", HttpStatus.INTERNAL_SERVER_ERROR);
+            throw new SimpleHttpException(400, "role naming conflict", HttpStatus.BAD_REQUEST);
         }
         return new RestfulResult(0, "role add successfully", new HashMap<>());
     }
@@ -71,8 +77,11 @@ public class AuthRoleController {
      */
     @RequestMapping(value = "/", method = RequestMethod.DELETE, produces = {MediaType.APPLICATION_JSON_VALUE, MediaType.APPLICATION_XML_VALUE})
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteRole(@RequestBody JSONObject input) throws ValidationException {
+    public void deleteRole(@RequestBody JSONObject input) throws ValidationException, SimpleHttpException {
         RBACPermissionRole permissionRole = ValidationUtils.validate(input.toJavaObject(RBACPermissionRole.class), NotNullRoleIdValidation.class);
+        if (permissionRole.getRoleId().equals(1)) {
+            throw new SimpleHttpException(400, "the default admin user can not be deleted", HttpStatus.BAD_REQUEST);
+        }
         permissionRoleService.deleteRole(permissionRole);
     }
 
@@ -83,6 +92,9 @@ public class AuthRoleController {
     @ResponseStatus(HttpStatus.CREATED)
     public RestfulResult renameRole(@RequestBody JSONObject input) throws ValidationException, SimpleHttpException {
         RBACPermissionRole permissionRole = ValidationUtils.validate(input.toJavaObject(RBACPermissionRole.class), NotNullRoleIdValidation.class, RoleOperationValidation.class);
+        if (permissionRole.getRoleId().equals(1)) {
+            throw new SimpleHttpException(400, "the default admin user can not be deleted", HttpStatus.BAD_REQUEST);
+        }
         if (!permissionRoleService.renameRole(permissionRole)) {
             throw new SimpleHttpException(500, "database access error", HttpStatus.INTERNAL_SERVER_ERROR);
         }
